adds asset_only lineage logic
